test(web): rename SoftwarePatternsSelection tests
