+++ web/src/components/software/SoftwarePage.test.jsx
@@ -25,7 +25,7 @@ import { act, screen, within } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { BUSY, IDLE } from "~/client/status";
 import { createClient } from "~/client";
-import test_patterns from "./PatternSelector.test.json";
+import test_patterns from "./SoftwarePatternsSelection.test.json";
 import SoftwarePage from "./SoftwarePage";
 
 jest.mock("~/client");
