+++ web/src/components/software/SoftwarePatternsSelection.test.json
