+++ web/src/components/software/SoftwarePatternsSelection.test.jsx
@@ -23,15 +23,15 @@ import React from "react";
 import { screen, within } from "@testing-library/react";
 import { plainRender } from "~/test-utils";
 
-import test_patterns from "./PatternSelector.test.json";
-import PatternSelector from "./PatternSelector";
+import test_patterns from "./SoftwarePatternsSelection.test.json";
+import SoftwarePatternsSelection from "./SoftwarePatternsSelection";
 import { SelectedBy } from "~/client/software";
 
 const patterns = test_patterns.map((p) => ({ ...p, selectedBy: SelectedBy.NONE }));
 
-describe.skip("PatternSelector", () => {
+describe.skip("SoftwarePatternsSelection", () => {
   it("displays the pattern groups in the correct order", () => {
-    plainRender(<PatternSelector patterns={patterns} />);
+    plainRender(<SoftwarePatternsSelection patterns={patterns} />);
     const headings = screen.getAllByRole("heading", { level: 2 });
     const headingsText = headings.map((node) => node.textContent);
     expect(headingsText).toEqual([
@@ -43,7 +43,7 @@ describe.skip("PatternSelector", () => {
 
   it("displays the patterns in a group in correct order", async () => {
     plainRender(
-      <PatternSelector patterns={patterns} />,
+      <SoftwarePatternsSelection patterns={patterns} />,
     );
 
     // the "Base Technologies" pattern group
@@ -59,7 +59,7 @@ describe.skip("PatternSelector", () => {
 
   it("displays only the matching patterns when using the search filter", async () => {
     const { user } = plainRender(
-      <PatternSelector patterns={patterns} />,
+      <SoftwarePatternsSelection patterns={patterns} />,
     );
 
     // enter "multimedia" into the search filter
@@ -81,7 +81,7 @@ describe.skip("PatternSelector", () => {
     pattern.selectedBy = SelectedBy.USER;
 
     plainRender(
-      <PatternSelector patterns={patterns} />,
+      <SoftwarePatternsSelection patterns={patterns} />,
     );
 
     // the "Base Technologies" pattern group
